+++ CHANGELOG.md
@@ -102,3 +102,4 @@
 * Update middleware
 * Add `ResponseAssertion`
 * Add CORS validation
+* Add toJSON setter in `DModel`